ezgripper-gui/backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
cyclonedds
unitree_sdk2py
